feat: create a dedicated api route for admin
